Migrate linter to use golangci-lint
This currently uses gometalinter which has differences in performance
than golangci-lint. It 'calls the linters directly and reuses work by
parsing program once'. From initial investigation it *is* faster, tried
on various projects but looks like they also need to do quite a bit more
work with program load times as that's still where the bottleneck it's
the majority of the time spent.

===

Don't directly replace gometalinter.

Following feedback I have re-introduced gometalinter as the default so
it's as backwards compatible as possible. I've also switched
golangci-lint to install using the script recommended by the author,
tied to the latest version which can be controlled by the environment
variable $GOLANGCI_LINT_VERSION.

===

Revert lint script

We don't want any of the changes I made to the lint script until I get
some feedback around whether we should allow the golangci-lint to be
used via flags or whether it should just be run manually, or part of a
different script.

Signed-off-by: Chris Mills <millscj01@gmail.com>